requirements.txt
hikari
requests
flask
firebase-admin
json-with-comments
roblox
hikari-lightbulb
python-dotenv
asyncio